requirements.txt
# requirements.txt
langgraph
langchain
langchain-openai
tavily-python
fastapi
uvicorn
pydantic
aiohttp
python-dotenv
numpy
scikit-learn
pymongo